Windows: Address more todos
